damage, health and slowdown into doubles

+++ src/main/scala/Logic/Bullet.scala
@@ -5,7 +5,7 @@ import _root_.Logic.Logic.GameObject
 import scalafx.geometry.Point2D
 
 // Target: add adjustable target????
-class Bullet (fileLoc: String, val target: Point2D, speed: Int, val slowDown: Int, val damage: Int, val targetEnemy: Enemy) extends GameObject(fileLoc, 35) {
+class Bullet (fileLoc: String, val target: Point2D, speed: Int, val slowDown: Double, val damage: Double, val targetEnemy: Enemy) extends GameObject(fileLoc, 35) {
     def move() = {
         val distToTarget = getGlobalCenter.distance(target)
         if (distToTarget > speed) {
